fix: handle bad `ctof_tdcadc` fits
If the first peak fit is bad, it's possible that the 2nd peak finder
will try to search for a peak in an _empty_ dataset, since the 1st
peak's sigma may be wider than the full histogram range. In this bad
case, just equate the 1st and 2nd peak positions so we get 2 bad fits
that are equivalent.