fix: `rich_dt_m` sigma bounds and timeout bad fits
RG-D run 18355 was stalling, because of a bad fit. I found that setting
the lower bound of `sigma` helps make this fit converge. I also added a
timeout to the fit routine, which will abort the fit after 10 seconds.

However, it seems that RG-D's fits are quite bad for this timeline, but
that's a problem for another PR.