refactor!: remove `plot` methods from monitoring classes
Plot viewing should use the web front-end, not custom canvases and PNG
output; we can allow webserver access to those who need it.

Similar to the removal of `main` methods, if someone depends on `plot`
methods, we may restore them from `git` history. We remove unused code
so that we don't have to waste time maintaining it.