hotfix: remove bad timeline HIPO files and allow Step 2 to finish
RG-L may start causing some timeline HIPO files to be bad/empty. In
particular, `ftof_ctof_vtdiff` is empty for current RG-L data.

This should not prevent Step 2 from running to completion, but we need
to prevent the bad HIPO files from being deployed. The bad timemlines
are now removed and the user is warned.