refactor: run-number dependent conditionals should use `RunDependentCut`
We have a lot of run-number dependence throughout, which is currently
tracked by #74. Instead of continuing to add to the list, in desperate
hope of a general solution, let's mitigate by keeping track of _where_
the run-number dependent conditionals are used. We can do this by using
custom run-number cut functions in a new class `RunDependentCut`.

We can't enforce its usage, but we can try to be consistent.

Many of these functions are very simple, but it makes it much easier to
find all the run-dependent cuts, should we return to the idea of some
configuration file implementation later.

I think this is good enough for now, since it's more flexible than a
configuration file, so I'd say this closes #74.