feat: add `--clobber` option to `qtl analysis`
This allows the target publishing directory to be _clobbered_, as
opposed to `--overwrite` (which removes and recreates).

@mathieuouillon reports that `--overwrite` has been causing the RG-L
auto-pass0 timeline link to temporarily disappear from the clas12mon
webpage; `--clobber` hopefully prevents that from happening.

Note that `--overwrite` will be needed for the rare case when we
_remove_ a timeline, so that they get removed from the publishing
directory (or just remove the HIPO files manually).